% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretodens.R
\name{dpareto1}
\alias{dpareto1}
\title{dpareto1 function for pareto distribution}
\usage{
dpareto1(x, alpha, x_m, log = FALSE)
}
\arguments{
\item{x}{x}

\item{alpha}{alpha}

\item{x_m}{x_m}

\item{log}{log}
}
\description{
dpareto1 function for pareto distribution is developed.
}
\examples{
dpareto1(1, 1, 1, log=TRUE)

}
